refactor(terminal): give Ghostty spike its own text surface
